+++ source/AA2_02_Arkanoid/Brick/HeavyBrick.h
@@ -5,7 +5,7 @@
 class HeavyBrick : public Brick {
 
 public:
-	HeavyBrick(SDL_Renderer* renderer, const Vector2D<int>& position, const Vector2D<int>& size, const int& spriteSourceWidthGap,
+	HeavyBrick(SDL_Renderer* renderer, const Vector2D<float>& position, const Vector2D<int>& size, const int& spriteSourceWidthGap,
 		const int& points, const int& lives);
 	~HeavyBrick();
