+++ source/AA2_02_Arkanoid/Brick/NormalBrick.h
@@ -5,7 +5,7 @@
 class NormalBrick : public Brick {
 
 public:
-	NormalBrick(SDL_Renderer* renderer, const Vector2D<int>& position, const Vector2D<int>& size, const int& spriteSourceWidthGap,
+	NormalBrick(SDL_Renderer* renderer, const Vector2D<float>& position, const Vector2D<int>& size, const int& spriteSourceWidthGap,
 		const int& points, const int& lives);
 	~NormalBrick();
